pyproject.toml
[project]
name = "fit-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.121.1",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "jinja2>=3.1.6",
    "pydantic-ai-slim[anthropic,logfire,openai]>=1.15.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]
